+++ test/unittests/test_cpp.cpp
@@ -2,6 +2,10 @@
 // Copyright 2018-2019 The EVMC Authors.
 // Licensed under the Apache License, Version 2.0.
 
+// The vector is not used here, but including it was causing compilation issues
+// previously related to using explicit template argument (SFINAE disabled).
+#include <vector>
+
 #include "../../examples/example_host.h"
 #include "../../examples/example_vm.h"
 
@@ -119,4 +123,78 @@ TEST(cpp, host_call)
     EXPECT_TRUE(std::equal(&res.output_data[0], &res.output_data[res.output_size], msg.input_data));
 
     example_host_destroy_context(host_context);
+}
+
+TEST(cpp, result_raii)
+{
+    static auto release_called = 0;
+    release_called = 0;
+    auto release_fn = [](const evmc_result*) noexcept { ++release_called; };
+
+    {
+        auto raw_result = evmc_result{};
+        raw_result.status_code = EVMC_INTERNAL_ERROR;
+        raw_result.release = release_fn;
+
+        auto raii_result = evmc::result{raw_result};
+        EXPECT_EQ(raii_result.status_code, EVMC_INTERNAL_ERROR);
+        EXPECT_EQ(raii_result.gas_left, 0);
+        raii_result.gas_left = -1;
+
+        auto raw_result2 = raii_result.raw();
+        EXPECT_EQ(raw_result2.status_code, EVMC_INTERNAL_ERROR);
+        EXPECT_EQ(raw_result.status_code, EVMC_INTERNAL_ERROR);
+        EXPECT_EQ(raw_result2.gas_left, -1);
+        EXPECT_EQ(raw_result.gas_left, 0);
+        EXPECT_EQ(raw_result2.release, release_fn);
+        EXPECT_EQ(raw_result.release, release_fn);
+    }
+    EXPECT_EQ(release_called, 0);
+
+    {
+        auto raw_result = evmc_result{};
+        raw_result.status_code = EVMC_INTERNAL_ERROR;
+        raw_result.release = release_fn;
+
+        auto raii_result = evmc::result{raw_result};
+        EXPECT_EQ(raii_result.status_code, EVMC_INTERNAL_ERROR);
+    }
+    EXPECT_EQ(release_called, 1);
+}
+
+TEST(cpp, result_move)
+{
+    static auto release_called = 0;
+    auto release_fn = [](const evmc_result*) noexcept { ++release_called; };
+
+    release_called = 0;
+    {
+        auto raw = evmc_result{};
+        raw.gas_left = -1;
+        raw.release = release_fn;
+
+        auto r0 = evmc::result{raw};
+        EXPECT_EQ(r0.gas_left, raw.gas_left);
+
+        auto r1 = std::move(r0);
+        EXPECT_EQ(r1.gas_left, raw.gas_left);
+    }
+    EXPECT_EQ(release_called, 1);
+
+    release_called = 0;
+    {
+        auto raw1 = evmc_result{};
+        raw1.gas_left = 1;
+        raw1.release = release_fn;
+
+        auto raw2 = evmc_result{};
+        raw2.gas_left = 1;
+        raw2.release = release_fn;
+
+        auto r1 = evmc::result{raw1};
+        auto r2 = evmc::result{raw2};
+
+        r2 = std::move(r1);
+    }
+    EXPECT_EQ(release_called, 2);
 }